\docType{package}
\name{fastLISA-package}
\alias{fastLISA}
\alias{fastLISA-package}
\title{fastLISA: Fast Local Indicators of Spatial Association}
\description{
Computes Local Indicators of Spatial Association (LISA) statistics using a
plain-C backend with optional OpenMP multi-threading and a permutation-based
significance test. See the package functions \code{\link{local_moran}},
\code{\link{local_moran_bv}}, \code{\link{local_moran_eb}},
\code{\link{local_geary}}, \code{\link{local_multigeary}},
\code{\link{local_g}}, and \code{\link{local_gstar}}.
}
\author{
\strong{Maintainer}: Lizhong Chen \email{chen.l@wehi.edu.au}

Authors:
\itemize{
  \item Lizhong Chen \email{chen.l@wehi.edu.au}
}

}
\keyword{internal}
